Moving context setup function to renderer init This is to facilitate swapchain recreation on window resize
