+++ src/Morrigu/Renderer/APIs/Vulkan/Helper.cpp
@@ -1,0 +1,27 @@
+#include "Helper.h"
+
+namespace MRG::Vulkan
+{
+	[[nodiscard]] SwapChainSupportDetails querySwapChainSupport(const VkPhysicalDevice device, const VkSurfaceKHR surface)
+	{
+		SwapChainSupportDetails details;
+		vkGetPhysicalDeviceSurfaceCapabilitiesKHR(device, surface, &details.capabilities);
+
+		uint32_t count{};
+
+		vkGetPhysicalDeviceSurfaceFormatsKHR(device, surface, &count, nullptr);
+		if (count != 0) {
+			details.formats.resize(count);
+			vkGetPhysicalDeviceSurfaceFormatsKHR(device, surface, &count, details.formats.data());
+		}
+
+		count = 0;
+		vkGetPhysicalDeviceSurfacePresentModesKHR(device, surface, &count, nullptr);
+		if (count != 0) {
+			details.presentModes.resize(count);
+			vkGetPhysicalDeviceSurfacePresentModesKHR(device, surface, &count, details.presentModes.data());
+		}
+
+		return details;
+	}
+}  // namespace MRG::Vulkan
